fix all test #5
