--- spec/mailers/enrollments_mailer_spec.rb
@@ -1,5 +1,0 @@
-require "spec_helper"
-
-describe EnrollmentsMailer do
-  pending "add some examples to (or delete) #{__FILE__}"
-end
