+++ spec/spec_helper.rb
@@ -3,30 +3,13 @@ ENV["RAILS_ENV"] ||= 'test'
 require File.expand_path("../../config/environment", __FILE__)
 require 'rspec/rails'
 require 'rspec/autorun'
-require "capybara/rspec"
-require "capybara/rails"
-require 'database_cleaner'
-require 'capybara/firebug'
+require 'capybara/rspec'
+require 'capybara/poltergeist'
 
-# Requires supporting ruby files with custom matchers and macros, etc,
-# in spec/support/ and its subdirectories.
-Dir[Rails.root.join("spec/support/**/*.rb")].each {|f| require f}
+Dir[Rails.root.join("spec/support/**/*.rb")].each { |f| require f }
+Capybara.javascript_driver = :poltergeist
 
 RSpec.configure do |config|
-  config.use_transactional_fixtures = false
-  # ## Mock Framework
-  #
-  # If you prefer to use mocha, flexmock or RR, uncomment the appropriate line:
-  #
-  # config.mock_with :mocha
-  # config.mock_with :flexmock
-  # config.mock_with :rr
-
-  # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
-  config.fixture_path = "#{::Rails.root}/spec/fixtures"
-  config.include(HelperMethods)
-
-  # Database Cleaner
   config.before(:suite) do
     DatabaseCleaner.strategy = :truncation
     DatabaseCleaner.clean_with(:truncation)
@@ -39,16 +22,7 @@ RSpec.configure do |config|
   config.after(:each) do
     DatabaseCleaner.clean
   end
-
-  # If you're not using ActiveRecord, or you'd prefer not to run each of your
-  # examples within a transaction, remove the following line or assign false
-  # instead of true.
-
-  # config.use_transactional_fixtures = true
-
-  # If true, the base class of anonymous controllers will be inferred
-  # automatically. This will be the default behavior in future versions of
-  # rspec-rails.
-  Capybara.javascript_driver = :selenium
-  Capybara.default_wait_time = 5
+  config.include(LoginMacros)
+  config.infer_base_class_for_anonymous_controllers = false
+  config.order = "random"
 end
