+++ spec/support/login_macros.rb
@@ -1,0 +1,8 @@
+module LoginMacros
+  def login(admin)
+    visit posts_path
+    fill_in 'admin_email', :with => admin.email
+    fill_in 'admin_password', :with => admin.password
+    click_button 'Acceder'
+  end
+end
